+++ scaleph-ui-react/config/routes.ts
@@ -144,14 +144,6 @@ export default [
             pCode: PRIVILEGE_CODE.datadevDatasourceShow,
             access: 'normalRouteFilter',
           },
-          {
-            name: 'batch',
-            path: '/workspace/dev/batch',
-            exact: true,
-            component: './DEV/DevBatchJob',
-            pCode: PRIVILEGE_CODE.datadevDatasourceShow,
-            access: 'normalRouteFilter',
-          },
         ],
       },
     ],
