--- scaleph-ui-react/src/pages/DEV/DevBatchJob/components/BasicForm.tsx
@@ -1,17 +1,0 @@
-import {ProCard, ProForm, ProFormSelect, ProFormText} from "@ant-design/pro-components";
-import {useIntl} from "@@/exports";
-import {ResourceListParam} from "@/services/resource/typings";
-import {listDictDataByType} from "@/services/admin/dictData.service";
-import {DICT_TYPE, RESOURCE_TYPE} from "@/constant";
-import {list} from "@/services/resource/resource.service";
-
-const BasicForm: React.FC = () => {
-  const intl = useIntl();
-
-  // [key:string]:any
-  return (<div>
-
-  </div>)
-}
-
-export default BasicForm;
